some correct, rename variables
