+++ docs/src/components/examples/Stripe/CompanyContent.vue
@@ -9,8 +9,8 @@
 </template>
 
 <script>
-import BaseGroup from '../../base/Group.vue';
-import BaseText from '../../base/Text.vue';
+import BaseGroup from '../../base/BaseGroup.vue';
+import BaseText from '../../base/BaseText.vue';
 
 export default {
   components: {
