+++ docs/src/components/examples/Stripe/MobileContent.vue
@@ -10,8 +10,8 @@
 </template>
 
 <script>
-import BaseGroup from '../../base/Group.vue';
-import BaseText from '../../base/Text.vue';
+import BaseGroup from '../../base/BaseGroup.vue';
+import BaseText from '../../base/BaseText.vue';
 
 export default {
   components: {
