+++ package.json
@@ -56,46 +56,46 @@
   },
   "devDependencies": {
     "@babel/core": "7.16.7",
-    "@babel/preset-env": "^7.18.2",
-    "@babel/preset-typescript": "^7.17.12",
-    "@types/jest": "^27.5.1",
-    "@types/node": "^17.0.35",
-    "@types/sinon": "^10.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.26.0",
-    "@typescript-eslint/parser": "^5.26.0",
-    "@vitejs/plugin-vue": "^2.3.3",
-    "@vue/compiler-sfc": "^3.2.36",
+    "@babel/preset-env": "^7.19.3",
+    "@babel/preset-typescript": "^7.18.6",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^17.0.45",
+    "@types/sinon": "^10.0.13",
+    "@typescript-eslint/eslint-plugin": "^5.38.1",
+    "@typescript-eslint/parser": "^5.38.1",
+    "@vitejs/plugin-vue": "^2.3.4",
+    "@vue/compiler-sfc": "^3.2.40",
     "@vue/eslint-config-prettier": "^7.0.0",
     "@vue/eslint-config-typescript": "^10.0.0",
-    "@vue/test-utils": "^2.0.0",
-    "autoprefixer": "^10.4.7",
+    "@vue/test-utils": "^2.1.0",
+    "autoprefixer": "^10.4.12",
     "babel-jest": "26.6.3",
-    "eslint": "^8.16.0",
+    "eslint": "^8.24.0",
     "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-vue": "^8.7.1",
-    "highlight.js": "^11.5.1",
+    "highlight.js": "^11.6.0",
     "jest": "^26.6.3",
-    "postcss": "^8.4.14",
-    "postcss-html": "^1.4.1",
-    "prettier": "^2.6.2",
-    "sass": "^1.52.1",
+    "postcss": "^8.4.17",
+    "postcss-html": "^1.5.0",
+    "prettier": "^2.7.1",
+    "sass": "^1.55.0",
     "sinon": "^10.0.0",
-    "stylelint": "^14.8.5",
+    "stylelint": "^14.13.0",
     "stylelint-config-prettier-scss": "^0.0.1",
     "stylelint-config-recommended-vue": "^1.4.0",
     "stylelint-config-sass-guidelines": "^9.0.1",
     "stylelint-config-standard-scss": "^3.0.0",
     "ts-jest": "^26.5.6",
-    "ts-node": "^10.8.0",
-    "typescript": "^4.7.2",
-    "vite": "^2.9.9",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.4",
+    "vite": "^2.9.15",
     "vue-eslint-parser": "^8.3.0",
     "vue-jest": "^5.0.0-alpha.10",
-    "webpack": "^5.72.1"
+    "webpack": "^5.74.0"
   },
   "dependencies": {
-    "vue": "^3.2.36"
+    "vue": "^3.2.40"
   }
 }
